Spike: Pend tests that 'DiskLimitScopeTotal', fix test for ext4 overhead
The ext4 FS on our new quota layer takes up a bunch of space, so when
writing a file that we expect to fit into the layer, we now have to make
it a bit smaller. We have no hope of passing tests that use
DiskLimitScopeTotal, since we currently only implement Exclusive quotas
in this spike.

Signed-off-by: Gareth Smith <gsmith@pivotal.io>