Pend cgroup test since it does not work non-nested
This test works for all nested environments (nested GATS or non-nested
GATS agains Guardian that is deployed in BOSH lite). However it does not
work for an actual BOSH deployment.

It seems that runC inherits the cgroups used by the host process. BOSH
lite and Concourse use the cpuset cgroup (since they use containers
based on garden-linux). However the BOSH stemcell
bosh-aws-xen-hvm-ubuntu-trusty-go_agent does not use the cpuset cgroup.
Hence this test is failing since it looks for it.